requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
frozenlist==1.3.3
gunicorn==20.0.4
idna==3.4
itsdangerous==1.1.0
Jinja2==2.11.3
load-dotenv==0.1.0
MarkupSafe==1.1.1
multidict==6.0.4
openai==0.27.8
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
tqdm==4.65.0
urllib3==2.0.3
Werkzeug==1.0.1
yarl==1.9.2
